chore(quest-preview): show money reward
