+++ src/app/features/quest/quest-preview/quest-preview.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { QuestPreviewService } from './quest-preview.service';
+import { PreviewHelperService } from '@keira-shared/services/preview-helper.service';
 
 @Component({
   selector: 'keira-quest-preview',
@@ -7,7 +8,11 @@ import { QuestPreviewService } from './quest-preview.service';
   styleUrls: ['./quest-preview.component.scss']
 })
 export class QuestPreviewComponent implements OnInit {
-  constructor(public readonly service: QuestPreviewService) { }
+
+  constructor(
+    public readonly service: QuestPreviewService,
+    public readonly helper: PreviewHelperService,
+  ) { }
 
   descriptionToggle = true;
   progressToggle = true;
